requirements.txt
altair==6.0.0
aspose_email_for_python_via_net==25.11
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
libpff-python==20231205
MarkupSafe==3.0.3
narwhals==2.14.0
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
pyarrow==22.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.52.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
watchdog==6.0.0
Whoosh==2.7.4
sentence-transformers==2.2.2
huggingface-hub==0.17.3
PyMuPDF==1.23.26
python-docx==1.1.0
chromadb==0.4.24
